Promote daObjKinokoTag_c genuine C++ TU
