--- src/_ZN16daObjKinokoTag_cD0Ev.cpp
@@ -1,8 +1,0 @@
-//cpp
-// @symbol _ZN16daObjKinokoTag_cD0Ev
-
-#include "daObjKinokoTag_c.h"
-
-daObjKinokoTag_c::~daObjKinokoTag_c()
-{
-}
